Move predicates into separate folder
